added api response codes and made encryption optional
I added an output for the Post function to tell you what the response code is from an API call so you can visually track failures. Not currently acting on those calls. I also made it optional to use the encryption files so you can still hardcode your credentials or wait to be prompted